Fix: compile error

+++ srcs/check_path.cpp
@@ -17,7 +17,7 @@ bool	Map::checkReached(int **visited)
 
 void	Map::find(int **visited, int y, int x)
 {
-	if (y < 0 || y >= this->y_size || x < 0 || x >= this->x_size)
+	if (y < 0 || y >= (int)this->y_size || x < 0 || x >= (int)this->x_size)
 		return ;
 	if (this->isEqualChar(y, x, '1'))
 		return ;
@@ -35,14 +35,14 @@ void	Map::find(int **visited, int y, int x)
 void	Map::isValidPath()
 {	
 	int** visited = new int*[this->y_size];
-	for (int i = 0; i < this->y_size; i++)
+	for (int i = 0; i < (int)this->y_size; i++)
 		visited[i] = new int[this->x_size];
 	
 	this->find(visited, this->start_y, this->start_x);
 	if (!this->checkReached(visited))
 		this->printError(INVALID_PATH);
 	
-	for(int i = 0; i < this->y_size; i++)
+	for(int i = 0; i < (int)this->y_size; i++)
 		delete[] visited[i];
 	delete[] visited;
 }
